Fixes premature search results bug
The initial "looking for titanium" for-loop was prematurly breaking
before any chance to find the correct location. It also attempted to
search all user directories. The side effect was that it would only look
in a user directory based on file globbing (presumably aphabetical) then
exit. The correct user directory could have been further down the list.
Also searching user directories other than your own is futile as you
don't have permissions to look in other people's home directory.

Also by default on Mac OS X if you install a titanium SDK via the
command line it stores it in your user directry. However if you use the
Titanium Studio it will install to the system. This means you can have
the correct sdk installed but if the search for it finds a different
titanium folder it would prematurly assume the wrong location.